fixup! refactor: platform billing controller auth guards
