+++ apps/api/v2/src/modules/billing/guards/is-user-in-billing-org.ts
@@ -30,7 +30,6 @@ export class IsUserInBillingOrg implements CanActivate {
       );
     }
 
-    request.user = user;
     return true;
   }
 }
